Estrapolate classi in file dedicati e usato il require per usali come prima tramite Tables.php
