+++ app/model/table/list/CatalogTable.php
@@ -1,0 +1,33 @@
+<?php declare(strict_types=1);
+
+	/**
+	 * Class representing the 'Catalogs' table.
+	 */
+	class CatalogTable
+	{
+		const TABLE_NAME = 'Catalogs';
+
+		const ID = 'id';
+		const NAME = 'name';
+		const DESCRIPTION = 'description';
+		const NOTE = 'note';
+		const CREATION_DATE = 'creationDate';
+		const UPDATE_DATE = 'updateDate';
+
+		/**
+		 * Get the list of columns for the Catalogs table.
+		 *
+		 * @return array List of column names
+		 */
+		public static function getColumns(): array
+		{
+			return [
+				self::ID,
+				self::NAME,
+				self::DESCRIPTION,
+				self::NOTE,
+				self::CREATION_DATE,
+				self::UPDATE_DATE
+			];
+		}
+	}
