+++ app/model/table/list/InfoProfileTypesTable.php
@@ -1,0 +1,37 @@
+<?php declare(strict_types=1);
+
+	/**
+	 * Class InfoProfileTypes
+	 *
+	 * Represents the structure and columns of the InfoProfileTypes table.
+	 */
+	class InfoProfileTypesTable
+	{
+		const TABLE_NAME = 'InfoProfileTypes';
+
+		const ID = 'id';
+		const PLAN = 'plan';
+		const IDPROFILETYPES = 'idProfileTypes';
+		const DESCRIPTION = 'description';
+		const NOTES = 'notes';
+		const CREATION_DATE = 'creation_date';
+		const UPDATE_DATE = 'update_date';
+
+		/**
+		 * Get the list of columns for the translations table.
+		 *
+		 * @return array List of column names.
+		 */
+		public static function getColumns(): array
+		{
+			return [
+				self::ID,
+				self::PLAN,
+				self::IDPROFILETYPES,
+				self::DESCRIPTION,
+				self::NOTES,
+				self::CREATION_DATE,
+				self::UPDATE_DATE
+			];
+		}
+	}
