+++ app/model/table/list/PropertyTable.php
@@ -1,0 +1,45 @@
+<?php declare(strict_types=1);
+
+	/**
+	 * Class representing the 'Properties' table.
+	 */
+	class PropertyTable
+	{
+		const TABLE_NAME = 'Properties';
+
+		const ID = 'id';
+		const NAME = 'name';
+		const DESCRIPTION = 'description';
+		const PRICE = 'price';
+		const ADDRESS = 'address';
+		const ID_CITY = 'IdCity';
+		const ID_STATE = 'IdState';
+		const ID_PropertyTypes = 'IdPropertyTypes';
+		const ACTIVE = 'active';
+		const NOTES = 'notes';
+		const CREATION_DATE = 'creationDate';
+		const UPDATE_DATE = 'updateDate';
+
+		/**
+		 * Get the list of columns for the Properties table.
+		 *
+		 * @return array List of column names
+		 */
+		public static function getColumns(): array
+		{
+			return [
+				self::ID,
+				self::NAME,
+				self::DESCRIPTION,
+				self::PRICE,
+				self::ADDRESS,
+				self::ID_CITY,
+				self::ID_STATE,
+				self::ID_PropertyTypes,
+				self::ACTIVE,
+				self::NOTES,
+				self::CREATION_DATE,
+				self::UPDATE_DATE
+			];
+		}
+	}
