+++ app/model/table/list/PropertyTypesCategoriesTable.php
@@ -1,0 +1,33 @@
+<?php declare(strict_types=1);
+
+	/**
+	 * Class PropertyTypesCategoriesTable
+	 *
+	 * This class defines the structure of the table 'lmgc_PropertyTypesCategories' with constants for columns.
+	 */
+	class PropertyTypesCategoriesTable
+	{
+		const TABLE_NAME = 'lmgc_PropertyTypesCategories';
+
+		const ID = 'id';
+		const CATEGORY = 'category';
+		const NOTES = 'notes';
+		const CREATE_DATE = 'createDate';
+		const UPDATE_DATE = 'updateDate';
+
+		/**
+		 * Get the list of columns for the PropertyTypesCategories table.
+		 *
+		 * @return array List of column names
+		 */
+		public static function getColumns(): array
+		{
+			return [
+				self::ID,
+				self::CATEGORY,
+				self::NOTES,
+				self::CREATE_DATE,
+				self::UPDATE_DATE
+			];
+		}
+	}
